+++ src/version4.rs
@@ -1,5 +1,6 @@
 #![cfg_attr(docsrs, doc(cfg(feature = "v4")))]
 
+use core::convert::TryFrom;
 use core::marker::PhantomData;
 
 use crate::common::{encode_b64, validate_footer_untrusted_token};
@@ -13,13 +14,12 @@ use crate::version::private::Version;
 use alloc::string::String;
 use alloc::vec::Vec;
 use blake2b::SecretKey as AuthKey;
-use ed25519_compact::{KeyPair, PublicKey, Signature};
+use ed25519_compact::{KeyPair, PublicKey, SecretKey, Signature};
 use orion::hazardous::mac::blake2b;
 use orion::hazardous::mac::blake2b::Blake2b;
 use orion::hazardous::stream::xchacha20;
 use xchacha20::Nonce as EncNonce;
 use xchacha20::SecretKey as EncKey;
-use zeroize::Zeroizing;
 
 #[derive(Debug, PartialEq, Clone)]
 /// Version 4 of the PASETO spec.
@@ -27,7 +27,7 @@ pub struct V4;
 
 impl Version for V4 {
     const LOCAL_KEY: usize = 32;
-    const SECRET_KEY: usize = 32;
+    const SECRET_KEY: usize = 32 + Self::PUBLIC_KEY; // Seed || PK
     const PUBLIC_KEY: usize = 32;
     const PUBLIC_SIG: usize = 64;
     const LOCAL_NONCE: usize = 32;
@@ -44,13 +44,27 @@ impl Version for V4 {
     }
 
     fn validate_secret_key(key_bytes: &[u8]) -> Result<(), Error> {
-        debug_assert_eq!(Self::LOCAL_KEY, Self::SECRET_KEY);
-        Self::validate_local_key(key_bytes)
+        if key_bytes.len() != Self::SECRET_KEY {
+            return Err(Error::Key);
+        }
+
+        Ok(())
     }
 
     fn validate_public_key(key_bytes: &[u8]) -> Result<(), Error> {
-        debug_assert_eq!(Self::LOCAL_KEY, Self::PUBLIC_KEY);
-        Self::validate_local_key(key_bytes)
+        if key_bytes.len() != Self::PUBLIC_KEY {
+            return Err(Error::Key);
+        }
+
+        Ok(())
+    }
+}
+
+impl TryFrom<&AsymmetricSecretKey<V4>> for AsymmetricPublicKey<V4> {
+    type Error = Error;
+
+    fn try_from(value: &AsymmetricSecretKey<V4>) -> Result<Self, Self::Error> {
+        AsymmetricPublicKey::<V4>::from(&value.as_bytes()[32..])
     }
 }
 
@@ -58,7 +72,7 @@ impl Generate<AsymmetricKeyPair<V4>, V4> for AsymmetricKeyPair<V4> {
     fn generate() -> Result<AsymmetricKeyPair<V4>, Error> {
         let key_pair = KeyPair::generate();
 
-        let secret = AsymmetricSecretKey::<V4>::from(&key_pair.sk[..32])
+        let secret = AsymmetricSecretKey::<V4>::from(key_pair.sk.as_ref())
             .map_err(|_| Error::KeyGeneration)?;
         let public = AsymmetricPublicKey::<V4>::from(key_pair.pk.as_ref())
             .map_err(|_| Error::KeyGeneration)?;
@@ -90,7 +104,6 @@ impl PublicToken {
     /// Create a public token.
     pub fn sign(
         secret_key: &AsymmetricSecretKey<V4>,
-        public_key: &AsymmetricPublicKey<V4>,
         message: &[u8],
         footer: Option<&[u8]>,
         implicit_assert: Option<&[u8]>,
@@ -99,15 +112,12 @@ impl PublicToken {
             return Err(Error::EmptyPayload);
         }
 
-        let mut raw_key = Zeroizing::new([0u8; 64]);
-        raw_key.as_mut()[..32].copy_from_slice(secret_key.as_bytes());
-        raw_key.as_mut()[32..].copy_from_slice(public_key.as_bytes());
-        let kp = KeyPair::from_slice(raw_key.as_ref()).map_err(|_| Error::Key)?;
+        let sk = SecretKey::from_slice(secret_key.as_bytes()).map_err(|_| Error::Key)?;
 
         let f = footer.unwrap_or(&[]);
         let i = implicit_assert.unwrap_or(&[]);
         let m2 = pae::pae(&[Self::HEADER.as_bytes(), message, f, i])?;
-        let sig = kp.sk.sign(m2, None);
+        let sig = sk.sign(m2, None);
 
         let mut m_sig: Vec<u8> = Vec::from(message);
         m_sig.extend_from_slice(sig.as_ref());
@@ -373,7 +383,7 @@ mod test_vectors {
         debug_assert!(test.secret_key.is_some());
 
         let sk = AsymmetricSecretKey::<V4>::from(
-            &hex::decode(test.secret_key.as_ref().unwrap()).unwrap()[..32],
+            &hex::decode(test.secret_key.as_ref().unwrap()).unwrap(),
         )
         .unwrap();
         let pk = AsymmetricPublicKey::<V4>::from(
@@ -402,7 +412,7 @@ mod test_vectors {
         let message = test.payload.as_ref().unwrap().as_str().unwrap();
 
         let actual =
-            PublicToken::sign(&sk, &pk, message.as_bytes(), footer, Some(implicit_assert)).unwrap();
+            PublicToken::sign(&sk, message.as_bytes(), footer, Some(implicit_assert)).unwrap();
         assert_eq!(actual, test.token, "Failed {:?}", test.name);
         let ut = UntrustedToken::<Public, V4>::try_from(&test.token).unwrap();
 
@@ -441,16 +451,16 @@ mod test_tokens {
     use crate::token::UntrustedToken;
     use core::convert::TryFrom;
 
-    // In version 2 tests, the SK used for public tokens is valid for the local as well.
-    // Not the case with version 4.
     const TEST_LOCAL_SK_BYTES: [u8; 32] = [
         112, 113, 114, 115, 116, 117, 118, 119, 120, 121, 122, 123, 124, 125, 126, 127, 128, 129,
         130, 131, 132, 133, 134, 135, 136, 137, 138, 139, 140, 141, 142, 143,
     ];
 
-    const TEST_SK_BYTES: [u8; 32] = [
+    const TEST_SK_BYTES: [u8; 64] = [
         180, 203, 251, 67, 223, 76, 226, 16, 114, 125, 149, 62, 74, 113, 51, 7, 250, 25, 187, 125,
-        159, 133, 4, 20, 56, 217, 225, 27, 148, 42, 55, 116,
+        159, 133, 4, 20, 56, 217, 225, 27, 148, 42, 55, 116, 30, 185, 219, 187, 188, 4, 124, 3,
+        253, 112, 96, 78, 0, 113, 240, 152, 126, 22, 178, 139, 117, 114, 37, 193, 31, 0, 65, 93,
+        14, 32, 177, 162,
     ];
 
     const TEST_PK_BYTES: [u8; 32] = [
@@ -468,8 +478,7 @@ mod test_tokens {
     fn test_gen_keypair() {
         let kp = AsymmetricKeyPair::<V4>::generate().unwrap();
 
-        let token =
-            PublicToken::sign(&kp.secret, &kp.public, MESSAGE.as_bytes(), None, None).unwrap();
+        let token = PublicToken::sign(&kp.secret, MESSAGE.as_bytes(), None, None).unwrap();
 
         let ut = UntrustedToken::<Public, V4>::try_from(&token).unwrap();
         assert!(PublicToken::verify(&kp.public, &ut, None, None).is_ok());
@@ -495,7 +504,6 @@ mod test_tokens {
         let kp = AsymmetricKeyPair::<V4>::generate().unwrap();
         let token = PublicToken::sign(
             &kp.secret,
-            &kp.public,
             MESSAGE.as_bytes(),
             Some(FOOTER.as_bytes()),
             None,
@@ -526,7 +534,7 @@ mod test_tokens {
         let test_sk = AsymmetricSecretKey::<V4>::from(&TEST_SK_BYTES).unwrap();
         let test_pk = AsymmetricPublicKey::<V4>::from(&TEST_PK_BYTES).unwrap();
 
-        let token = PublicToken::sign(&test_sk, &test_pk, MESSAGE.as_bytes(), None, None).unwrap();
+        let token = PublicToken::sign(&test_sk, MESSAGE.as_bytes(), None, None).unwrap();
         let ut = UntrustedToken::<Public, V4>::try_from(&token).unwrap();
 
         assert!(PublicToken::verify(&test_pk, &ut, None, None).is_ok());
@@ -534,7 +542,7 @@ mod test_tokens {
 
     #[test]
     fn footer_logic() {
-        let test_local_sk = SymmetricKey::<V4>::from(&TEST_SK_BYTES).unwrap();
+        let test_local_sk = SymmetricKey::<V4>::from(&TEST_LOCAL_SK_BYTES).unwrap();
         let test_sk = AsymmetricSecretKey::<V4>::from(&TEST_SK_BYTES).unwrap();
         let test_pk = AsymmetricPublicKey::<V4>::from(&TEST_PK_BYTES).unwrap();
         let message =
@@ -542,11 +550,11 @@ mod test_tokens {
 
         // We create a token with Some(footer) and with None
         let actual_some = UntrustedToken::<Public, V4>::try_from(
-            &PublicToken::sign(&test_sk, &test_pk, message, Some(FOOTER.as_bytes()), None).unwrap(),
+            &PublicToken::sign(&test_sk, message, Some(FOOTER.as_bytes()), None).unwrap(),
         )
         .unwrap();
         let actual_none = UntrustedToken::<Public, V4>::try_from(
-            &PublicToken::sign(&test_sk, &test_pk, message, None, None).unwrap(),
+            &PublicToken::sign(&test_sk, message, None, None).unwrap(),
         )
         .unwrap();
 
@@ -586,7 +594,7 @@ mod test_tokens {
 
     #[test]
     fn implicit_none_some_empty_is_same() {
-        let test_local_sk = SymmetricKey::<V4>::from(&TEST_SK_BYTES).unwrap();
+        let test_local_sk = SymmetricKey::<V4>::from(&TEST_LOCAL_SK_BYTES).unwrap();
         let test_sk = AsymmetricSecretKey::<V4>::from(&TEST_SK_BYTES).unwrap();
         let test_pk = AsymmetricPublicKey::<V4>::from(&TEST_PK_BYTES).unwrap();
         let message =
@@ -594,11 +602,11 @@ mod test_tokens {
         let implicit = b"";
 
         let actual_some = UntrustedToken::<Public, V4>::try_from(
-            &PublicToken::sign(&test_sk, &test_pk, message, None, Some(implicit)).unwrap(),
+            &PublicToken::sign(&test_sk, message, None, Some(implicit)).unwrap(),
         )
         .unwrap();
         let actual_none = UntrustedToken::<Public, V4>::try_from(
-            &PublicToken::sign(&test_sk, &test_pk, message, None, None).unwrap(),
+            &PublicToken::sign(&test_sk, message, None, None).unwrap(),
         )
         .unwrap();
         assert_eq!(actual_some, actual_none);
@@ -623,12 +631,11 @@ mod test_tokens {
     #[test]
     // NOTE: See https://github.com/paseto-standard/paseto-spec/issues/17
     fn empty_payload() {
-        let test_local_sk = SymmetricKey::<V4>::from(&TEST_SK_BYTES).unwrap();
+        let test_local_sk = SymmetricKey::<V4>::from(&TEST_LOCAL_SK_BYTES).unwrap();
         let test_sk = AsymmetricSecretKey::<V4>::from(&TEST_SK_BYTES).unwrap();
-        let test_pk = AsymmetricPublicKey::<V4>::from(&TEST_PK_BYTES).unwrap();
 
         assert_eq!(
-            PublicToken::sign(&test_sk, &test_pk, b"", None, None).unwrap_err(),
+            PublicToken::sign(&test_sk, b"", None, None).unwrap_err(),
             Error::EmptyPayload
         );
         assert_eq!(
@@ -920,9 +927,9 @@ mod test_keys {
 
     #[test]
     fn test_invalid_sizes() {
-        assert!(AsymmetricSecretKey::<V4>::from(&[0u8; 31]).is_err());
-        assert!(AsymmetricSecretKey::<V4>::from(&[0u8; 32]).is_ok());
-        assert!(AsymmetricSecretKey::<V4>::from(&[0u8; 33]).is_err());
+        assert!(AsymmetricSecretKey::<V4>::from(&[0u8; 63]).is_err());
+        assert!(AsymmetricSecretKey::<V4>::from(&[0u8; 64]).is_ok());
+        assert!(AsymmetricSecretKey::<V4>::from(&[0u8; 65]).is_err());
 
         assert!(AsymmetricPublicKey::<V4>::from(&[0u8; 31]).is_err());
         assert!(AsymmetricPublicKey::<V4>::from(&[0u8; 32]).is_ok());
@@ -931,6 +938,15 @@ mod test_keys {
         assert!(SymmetricKey::<V4>::from(&[0u8; 31]).is_err());
         assert!(SymmetricKey::<V4>::from(&[0u8; 32]).is_ok());
         assert!(SymmetricKey::<V4>::from(&[0u8; 33]).is_err());
+    }
+
+    #[test]
+    fn try_from_secret_to_public() {
+        let kpv4 = AsymmetricKeyPair::<V4>::generate().unwrap();
+        let pubv4 = AsymmetricPublicKey::<V4>::try_from(&kpv4.secret).unwrap();
+        assert_eq!(pubv4.as_bytes(), kpv4.public.as_bytes());
+        assert_eq!(pubv4, kpv4.public);
+        assert_eq!(&kpv4.secret.as_bytes()[32..], pubv4.as_bytes());
     }
 
     #[test]
@@ -964,13 +980,5 @@ impl AsymmetricKeyPair<V4> {
             secret: AsymmetricSecretKey::from(&bytes[..V4::SECRET_KEY])?,
             public: AsymmetricPublicKey::from(&bytes[V4::SECRET_KEY..])?,
         })
-    }
-
-    pub(crate) fn as_bytes<'a>(&self) -> [u8; 64] {
-        let mut buf = [0u8; V4::SECRET_KEY + V4::PUBLIC_KEY];
-        buf[..V4::SECRET_KEY].copy_from_slice(self.secret.as_bytes());
-        buf[V4::SECRET_KEY..].copy_from_slice(self.public.as_bytes());
-
-        buf
     }
 }